Move print_pipeline_data to nu-protocol
